move web scan loader tests
