+++ api-js/src/loaders/web-scan/__tests__/load-ssl-connections-by-domain-id.test.js
@@ -5,16 +5,13 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
-const {
-  httpsLoaderConnectionsByDomainId,
-  httpsLoaderByKey,
-} = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
+const { sslLoaderByKey, sslLoaderConnectionsByDomainId } = require('../..')
 
-describe('given the load https connection function', () => {
+describe('given the load ssl connection function', () => {
   let query, drop, truncate, migrate, collections, user, domain, i18n
 
   const consoleWarnOutput = []
@@ -64,49 +61,49 @@ describe('given the load https connection function', () => {
   })
 
   describe('given a successful load', () => {
-    let httpsScan1, httpsScan2
+    let sslScan1, sslScan2
     beforeEach(async () => {
-      httpsScan1 = await collections.https.save({
+      sslScan1 = await collections.ssl.save({
         timestamp: '2020-10-02T12:43:39Z',
       })
-      httpsScan2 = await collections.https.save({
+      sslScan2 = await collections.ssl.save({
         timestamp: '2020-10-03T12:43:39Z',
       })
-      await collections.domainsHTTPS.save({
-        _to: httpsScan1._id,
+      await collections.domainsSSL.save({
+        _to: sslScan1._id,
         _from: domain._id,
       })
-      await collections.domainsHTTPS.save({
-        _to: httpsScan2._id,
+      await collections.domainsSSL.save({
+        _to: sslScan2._id,
         _from: domain._id,
       })
     })
     describe('using no cursor', () => {
-      it('returns multiple https scans', async () => {
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+      it('returns multiple ssl scans', async () => {
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const httpsLoader = httpsLoaderByKey(query, i18n)
-        const expectedHttpsScans = await httpsLoader.loadMany([
-          httpsScan1._key,
-          httpsScan2._key,
+        const sslLoader = sslLoaderByKey(query, i18n)
+        const expectedSslScans = await sslLoader.loadMany([
+          sslScan1._key,
+          sslScan2._key,
         ])
 
-        expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-        expectedHttpsScans[0].domainId = domain._id
+        expectedSslScans[0].id = expectedSslScans[0]._key
+        expectedSslScans[0].domainId = domain._id
 
-        expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-        expectedHttpsScans[1].domainId = domain._id
+        expectedSslScans[1].id = expectedSslScans[1]._key
+        expectedSslScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -114,15 +111,15 @@ describe('given the load https connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[0]._key),
               node: {
-                ...expectedHttpsScans[0],
+                ...expectedSslScans[0],
               },
             },
             {
-              cursor: toGlobalId('https', expectedHttpsScans[1]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[1]._key),
               node: {
-                ...expectedHttpsScans[1],
+                ...expectedSslScans[1],
               },
             },
           ],
@@ -130,41 +127,41 @@ describe('given the load https connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: false,
-            startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-            endCursor: toGlobalId('https', expectedHttpsScans[1]._key),
+            startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+            endCursor: toGlobalId('ssl', expectedSslScans[1]._key),
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
     describe('using after cursor', () => {
-      it('returns https scan(s) after a given node id', async () => {
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+      it('returns ssl scan(s) after a given node id', async () => {
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const httpsLoader = httpsLoaderByKey(query, i18n)
-        const expectedHttpsScans = await httpsLoader.loadMany([
-          httpsScan1._key,
-          httpsScan2._key,
+        const sslLoader = sslLoaderByKey(query, i18n)
+        const expectedSslScans = await sslLoader.loadMany([
+          sslScan1._key,
+          sslScan2._key,
         ])
 
-        expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-        expectedHttpsScans[0].domainId = domain._id
+        expectedSslScans[0].id = expectedSslScans[0]._key
+        expectedSslScans[0].domainId = domain._id
 
-        expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-        expectedHttpsScans[1].domainId = domain._id
+        expectedSslScans[1].id = expectedSslScans[1]._key
+        expectedSslScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          after: toGlobalId('https', expectedHttpsScans[0]._key),
+          after: toGlobalId('ssl', expectedSslScans[0]._key),
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -172,9 +169,9 @@ describe('given the load https connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('https', expectedHttpsScans[1]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[1]._key),
               node: {
-                ...expectedHttpsScans[1],
+                ...expectedSslScans[1],
               },
             },
           ],
@@ -182,41 +179,41 @@ describe('given the load https connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('https', expectedHttpsScans[1]._key),
-            endCursor: toGlobalId('https', expectedHttpsScans[1]._key),
+            startCursor: toGlobalId('ssl', expectedSslScans[1]._key),
+            endCursor: toGlobalId('ssl', expectedSslScans[1]._key),
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
     describe('using before cursor', () => {
-      it('returns https scan(s) before a given node id', async () => {
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+      it('returns ssl scan(s) before a given node id', async () => {
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const httpsLoader = httpsLoaderByKey(query, i18n)
-        const expectedHttpsScans = await httpsLoader.loadMany([
-          httpsScan1._key,
-          httpsScan2._key,
+        const sslLoader = sslLoaderByKey(query, i18n)
+        const expectedSslScans = await sslLoader.loadMany([
+          sslScan1._key,
+          sslScan2._key,
         ])
 
-        expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-        expectedHttpsScans[0].domainId = domain._id
+        expectedSslScans[0].id = expectedSslScans[0]._key
+        expectedSslScans[0].domainId = domain._id
 
-        expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-        expectedHttpsScans[1].domainId = domain._id
+        expectedSslScans[1].id = expectedSslScans[1]._key
+        expectedSslScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          before: toGlobalId('https', expectedHttpsScans[1]._key),
+          before: toGlobalId('ssl', expectedSslScans[1]._key),
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -224,9 +221,9 @@ describe('given the load https connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[0]._key),
               node: {
-                ...expectedHttpsScans[0],
+                ...expectedSslScans[0],
               },
             },
           ],
@@ -234,40 +231,40 @@ describe('given the load https connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-            endCursor: toGlobalId('https', expectedHttpsScans[0]._key),
+            startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+            endCursor: toGlobalId('ssl', expectedSslScans[0]._key),
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
     describe('using first limit', () => {
-      it('returns the first n amount of https scan(s)', async () => {
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+      it('returns the first n amount of ssl scan(s)', async () => {
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const httpsLoader = httpsLoaderByKey(query, i18n)
-        const expectedHttpsScans = await httpsLoader.loadMany([
-          httpsScan1._key,
-          httpsScan2._key,
+        const sslLoader = sslLoaderByKey(query, i18n)
+        const expectedSslScans = await sslLoader.loadMany([
+          sslScan1._key,
+          sslScan2._key,
         ])
 
-        expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-        expectedHttpsScans[0].domainId = domain._id
+        expectedSslScans[0].id = expectedSslScans[0]._key
+        expectedSslScans[0].domainId = domain._id
 
-        expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-        expectedHttpsScans[1].domainId = domain._id
+        expectedSslScans[1].id = expectedSslScans[1]._key
+        expectedSslScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 1,
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -275,9 +272,9 @@ describe('given the load https connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[0]._key),
               node: {
-                ...expectedHttpsScans[0],
+                ...expectedSslScans[0],
               },
             },
           ],
@@ -285,40 +282,40 @@ describe('given the load https connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-            endCursor: toGlobalId('https', expectedHttpsScans[0]._key),
+            startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+            endCursor: toGlobalId('ssl', expectedSslScans[0]._key),
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
     describe('using last limit', () => {
-      it('returns the last n amount of https scan(s)', async () => {
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+      it('returns the last n amount of ssl scan(s)', async () => {
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const httpsLoader = httpsLoaderByKey(query, i18n)
-        const expectedHttpsScans = await httpsLoader.loadMany([
-          httpsScan1._key,
-          httpsScan2._key,
+        const sslLoader = sslLoaderByKey(query, i18n)
+        const expectedSslScans = await sslLoader.loadMany([
+          sslScan1._key,
+          sslScan2._key,
         ])
 
-        expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-        expectedHttpsScans[0].domainId = domain._id
+        expectedSslScans[0].id = expectedSslScans[0]._key
+        expectedSslScans[0].domainId = domain._id
 
-        expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-        expectedHttpsScans[1].domainId = domain._id
+        expectedSslScans[1].id = expectedSslScans[1]._key
+        expectedSslScans[1].domainId = domain._id
 
         const connectionArgs = {
           last: 1,
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -326,9 +323,9 @@ describe('given the load https connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('https', expectedHttpsScans[1]._key),
+              cursor: toGlobalId('ssl', expectedSslScans[1]._key),
               node: {
-                ...expectedHttpsScans[1],
+                ...expectedSslScans[1],
               },
             },
           ],
@@ -336,52 +333,52 @@ describe('given the load https connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('https', expectedHttpsScans[1]._key),
-            endCursor: toGlobalId('https', expectedHttpsScans[1]._key),
+            startCursor: toGlobalId('ssl', expectedSslScans[1]._key),
+            endCursor: toGlobalId('ssl', expectedSslScans[1]._key),
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
     describe('using date filter', () => {
-      let httpsScan3
+      let sslScan3
       beforeEach(async () => {
-        httpsScan3 = await collections.https.save({
+        sslScan3 = await collections.ssl.save({
           timestamp: '2020-10-04T12:43:39Z',
         })
-        await collections.domainsHTTPS.save({
-          _to: httpsScan3._id,
+        await collections.domainsSSL.save({
+          _to: sslScan3._id,
           _from: domain._id,
         })
       })
       describe('using start date filter', () => {
-        it('returns https scans at and after the start date', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+        it('returns ssl scans at and after the start date', async () => {
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const httpsLoader = httpsLoaderByKey(query, i18n)
-          const expectedHttpsScans = await httpsLoader.loadMany([
-            httpsScan2._key,
-            httpsScan3._key,
+          const sslLoader = sslLoaderByKey(query)
+          const expectedSslScans = await sslLoader.loadMany([
+            sslScan2._key,
+            sslScan3._key,
           ])
 
-          expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-          expectedHttpsScans[0].domainId = domain._id
+          expectedSslScans[0].id = expectedSslScans[0]._key
+          expectedSslScans[0].domainId = domain._id
 
-          expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-          expectedHttpsScans[1].domainId = domain._id
+          expectedSslScans[1].id = expectedSslScans[1]._key
+          expectedSslScans[1].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
             startDate: '2020-10-03',
           }
 
-          const httpsScans = await connectionLoader({
+          const sslScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -389,15 +386,15 @@ describe('given the load https connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+                cursor: toGlobalId('ssl', expectedSslScans[0]._key),
                 node: {
-                  ...expectedHttpsScans[0],
+                  ...expectedSslScans[0],
                 },
               },
               {
-                cursor: toGlobalId('https', expectedHttpsScans[1]._key),
+                cursor: toGlobalId('ssl', expectedSslScans[1]._key),
                 node: {
-                  ...expectedHttpsScans[1],
+                  ...expectedSslScans[1],
                 },
               },
             ],
@@ -405,41 +402,41 @@ describe('given the load https connection function', () => {
             pageInfo: {
               hasNextPage: false,
               hasPreviousPage: true,
-              startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-              endCursor: toGlobalId('https', expectedHttpsScans[1]._key),
+              startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+              endCursor: toGlobalId('ssl', expectedSslScans[1]._key),
             },
           }
 
-          expect(httpsScans).toEqual(expectedStructure)
+          expect(sslScans).toEqual(expectedStructure)
         })
       })
       describe('using end date filter', () => {
-        it('returns https scans at and before the end date', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+        it('returns ssl scans at and before the end date', async () => {
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const httpsLoader = await httpsLoaderByKey(query, i18n)
-          const expectedHttpsScans = await httpsLoader.loadMany([
-            httpsScan1._key,
-            httpsScan2._key,
+          const sslLoader = sslLoaderByKey(query, i18n)
+          const expectedSslScans = await sslLoader.loadMany([
+            sslScan1._key,
+            sslScan2._key,
           ])
 
-          expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-          expectedHttpsScans[0].domainId = domain._id
+          expectedSslScans[0].id = expectedSslScans[0]._key
+          expectedSslScans[0].domainId = domain._id
 
-          expectedHttpsScans[1].id = expectedHttpsScans[1]._key
-          expectedHttpsScans[1].domainId = domain._id
+          expectedSslScans[1].id = expectedSslScans[1]._key
+          expectedSslScans[1].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
             endDate: '2020-10-03T13:50:00Z',
           }
 
-          const httpsScans = await connectionLoader({
+          const sslScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -447,15 +444,15 @@ describe('given the load https connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+                cursor: toGlobalId('ssl', expectedSslScans[0]._key),
                 node: {
-                  ...expectedHttpsScans[0],
+                  ...expectedSslScans[0],
                 },
               },
               {
-                cursor: toGlobalId('https', expectedHttpsScans[1]._key),
+                cursor: toGlobalId('ssl', expectedSslScans[1]._key),
                 node: {
-                  ...expectedHttpsScans[1],
+                  ...expectedSslScans[1],
                 },
               },
             ],
@@ -463,30 +460,28 @@ describe('given the load https connection function', () => {
             pageInfo: {
               hasNextPage: true,
               hasPreviousPage: false,
-              startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-              endCursor: toGlobalId('https', expectedHttpsScans[1]._key),
+              startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+              endCursor: toGlobalId('ssl', expectedSslScans[1]._key),
             },
           }
 
-          expect(httpsScans).toEqual(expectedStructure)
+          expect(sslScans).toEqual(expectedStructure)
         })
       })
       describe('using start and end date filters', () => {
         it('returns a scan on a specific date', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const httpsLoader = httpsLoaderByKey(query, i18n)
-          const expectedHttpsScans = await httpsLoader.loadMany([
-            httpsScan2._key,
-          ])
+          const sslLoader = sslLoaderByKey(query, i18n)
+          const expectedSslScans = await sslLoader.loadMany([sslScan2._key])
 
-          expectedHttpsScans[0].id = expectedHttpsScans[0]._key
-          expectedHttpsScans[0].domainId = domain._id
+          expectedSslScans[0].id = expectedSslScans[0]._key
+          expectedSslScans[0].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
@@ -494,7 +489,7 @@ describe('given the load https connection function', () => {
             endDate: '2020-10-03T23:59:59Z',
           }
 
-          const httpsScans = await connectionLoader({
+          const sslScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -502,9 +497,9 @@ describe('given the load https connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('https', expectedHttpsScans[0]._key),
+                cursor: toGlobalId('ssl', expectedSslScans[0]._key),
                 node: {
-                  ...expectedHttpsScans[0],
+                  ...expectedSslScans[0],
                 },
               },
             ],
@@ -512,20 +507,20 @@ describe('given the load https connection function', () => {
             pageInfo: {
               hasNextPage: true,
               hasPreviousPage: true,
-              startCursor: toGlobalId('https', expectedHttpsScans[0]._key),
-              endCursor: toGlobalId('https', expectedHttpsScans[0]._key),
+              startCursor: toGlobalId('ssl', expectedSslScans[0]._key),
+              endCursor: toGlobalId('ssl', expectedSslScans[0]._key),
             },
           }
 
-          expect(httpsScans).toEqual(expectedStructure)
+          expect(sslScans).toEqual(expectedStructure)
         })
       })
     })
-    describe('no https scans are found', () => {
+    describe('no ssl scans are found', () => {
       it('returns an empty structure', async () => {
         await truncate()
 
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -536,7 +531,7 @@ describe('given the load https connection function', () => {
           first: 5,
         }
 
-        const httpsScans = await connectionLoader({
+        const sslScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -552,7 +547,7 @@ describe('given the load https connection function', () => {
           },
         }
 
-        expect(httpsScans).toEqual(expectedStructure)
+        expect(sslScans).toEqual(expectedStructure)
       })
     })
   })
@@ -571,7 +566,7 @@ describe('given the load https connection function', () => {
     describe('given an unsuccessful load', () => {
       describe('both limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -588,18 +583,19 @@ describe('given the load https connection function', () => {
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'You must provide a `first` or `last` value to properly paginate the `https` connection.',
+                'You must provide a `first` or `last` value to properly paginate the `ssl` connection.',
               ),
             )
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: httpsLoaderConnectionsByDomainId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: sslLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('first and last arguments are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -619,19 +615,20 @@ describe('given the load https connection function', () => {
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'Passing both `first` and `last` to paginate the `https` connection is not supported.',
+                'Passing both `first` and `last` to paginate the `ssl` connection is not supported.',
               ),
             )
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} tried to have \`first\` and \`last\` arguments set for: httpsLoaderConnectionsByDomainId.`,
+            `User: ${user._key} tried to have \`first\` and \`last\` arguments set for: sslLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('limits are below minimum', () => {
+      describe('both limits are below minimum', () => {
         describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -650,18 +647,19 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`first` on the `https` connection cannot be less than zero.',
+                  '`first` on the `ssl` connection cannot be less than zero.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -680,20 +678,21 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`last` on the `https` connection cannot be less than zero.',
+                  '`last` on the `ssl` connection cannot be less than zero.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
-      describe('limits are above maximum', () => {
+      describe('both limits are above maximum', () => {
         describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -712,18 +711,19 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 101 records on the `https` connection exceeds the `first` limit of 100 records.',
+                  'Requesting 101 records on the `ssl` connection exceeds the `first` limit of 100 records.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 101 for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set to 101 for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -742,12 +742,13 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 500 records on the `https` connection exceeds the `last` limit of 100 records.',
+                  'Requesting 500 records on the `ssl` connection exceeds the `last` limit of 100 records.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -758,7 +759,7 @@ describe('given the load https connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = httpsLoaderConnectionsByDomainId(
+              const connectionLoader = sslLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -784,7 +785,7 @@ describe('given the load https connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: httpsLoaderConnectionsByDomainId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: sslLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -794,7 +795,7 @@ describe('given the load https connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = httpsLoaderConnectionsByDomainId(
+              const connectionLoader = sslLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -820,7 +821,7 @@ describe('given the load https connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: httpsLoaderConnectionsByDomainId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: sslLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -833,7 +834,7 @@ describe('given the load https connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -851,12 +852,12 @@ describe('given the load https connection function', () => {
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load https scans. Please try again.'),
+            new Error('Unable to load ssl scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get https information for ${domain._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get ssl information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -869,7 +870,7 @@ describe('given the load https connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -887,12 +888,12 @@ describe('given the load https connection function', () => {
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load https scans. Please try again.'),
+            new Error('Unable to load ssl scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get https information for ${domain._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get ssl information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
@@ -912,7 +913,7 @@ describe('given the load https connection function', () => {
     describe('given an unsuccessful load', () => {
       describe('both limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -929,14 +930,15 @@ describe('given the load https connection function', () => {
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: httpsLoaderConnectionsByDomainId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: sslLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('first and last arguments are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = httpsLoaderConnectionsByDomainId(
+          const connectionLoader = sslLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -956,15 +958,16 @@ describe('given the load https connection function', () => {
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} tried to have \`first\` and \`last\` arguments set for: httpsLoaderConnectionsByDomainId.`,
+            `User: ${user._key} tried to have \`first\` and \`last\` arguments set for: sslLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('limits are below minimum', () => {
+      describe('both limits are below minimum', () => {
         describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -983,14 +986,15 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -1009,16 +1013,17 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
-      describe('limits are above maximum', () => {
+      describe('both limits are above maximum', () => {
         describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -1037,14 +1042,15 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 101 for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set to 101 for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = httpsLoaderConnectionsByDomainId(
+            const connectionLoader = sslLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -1063,8 +1069,9 @@ describe('given the load https connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: httpsLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: sslLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -1075,7 +1082,7 @@ describe('given the load https connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = httpsLoaderConnectionsByDomainId(
+              const connectionLoader = sslLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -1097,7 +1104,7 @@ describe('given the load https connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: httpsLoaderConnectionsByDomainId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: sslLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1107,7 +1114,7 @@ describe('given the load https connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = httpsLoaderConnectionsByDomainId(
+              const connectionLoader = sslLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -1129,7 +1136,7 @@ describe('given the load https connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: httpsLoaderConnectionsByDomainId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: sslLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1142,7 +1149,7 @@ describe('given the load https connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -1163,7 +1170,7 @@ describe('given the load https connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get https information for ${domain._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get ssl information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -1176,7 +1183,7 @@ describe('given the load https connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = httpsLoaderConnectionsByDomainId(
+        const connectionLoader = sslLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -1197,7 +1204,7 @@ describe('given the load https connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get https information for ${domain._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get ssl information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
